+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `ssh_upload()` and `ssh_download()` functions to upload/download files via SSH
 * Rename `ssh()` to `ssh_run()`
+* Allow to set default context with an env variable
 
 ## 0.9.1 (2023-10-09)